wx_django/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
idna==2.8
PyMySQL==0.9.3
python-Levenshtein==0.12.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
